Fix fatal bug
While looping to check if the command given is a builtin, the string in
the line it would check would be indexed the same as the builtin it
searched for.

For example, typing in:

exit

Would check for the second string in the line, which is nonexistent,
always giving a segmentation fault. Also happened with commands with no
arguments, as long as the number of builtins >1.

Fix that bug.